Revert "Add Optional class"
This reverts commit ef4166ad2b2375eb328acaf4c6b59e5cc813c6d1, and a few
later changes. The Optional wrapper wasn't very well though through, and
had some bugs. typescript is pretty good at forcing possible nulls to be
checked, so we can rely on this instead.

Fixes whtaylor/icat-admin#21, was caused by a call to get() on an object
that used to be an array, but had become an Optional.