Fix DataPublication related entity links
I found a very weird problem whilst working on this - the second test in 'context-menu.cy.js' fails, but only in firefox _as long as a breakpoint is not set in the debugger_. This was not introduced in this commit, I just happened to find it whilst experimenting with the new test in that file. I think it has something to do with the autoreloading the last open tab functionality, and possibly routing, which I'm planning to review soon, so am going to leave for now and hope that ironing out those parts fixes this problem.