Create an input wrapper that handles onchange events
React causes onChange attributes in JSX to bind the input event on the
DOM element. Now that we're using react-query, preact/compat has kicked
in and is matching that behaviour, but we still want the original
onChange behaviour for some inputs, which this component provides.

This component is based off of
https://stackoverflow.com/questions/38256332/in-react-whats-the-difference-between-onchange-and-oninput/62383569#62383569,
but ported to be a functional component, and with the ability to handle
a ref being passed in to it.