Fix/improve error handling
As part of updating to react-query v5 in dafd11b, I changed to using 'error.message' instead of just 'error', because react-query changed the type from string to Error. Unfortunately, the actual runtime type was string, so this broke the error handling, causing the app to crash on any error response from ICAT. To handle this correctly, we now resolve errors into an Error object, rather than a string, which seems to be the expected approach.